+++ README.md
@@ -49,32 +49,26 @@
 
 ## 회원 운동 기록
 ### 회원 별 운동 기록 제공
-- [ ] 회원은 자신의 운동 내역 기록할 수 있다.
+- [ ] 회원은 자신의 운동 내역을 기록할 수 있다.
   - [ ] 운동 내역은 세트별로 기록한다.
   - [ ] 기록한 내용은 날짜 별로 확인할 수 있다.
 - [ ] 운동 종목 별 최고 기록 갱신 시 알림 기능 `sse/polling/websocket 등`
-- [ ] 하루 운동이 끝나면 `가슴/등/이두/삼두/어깨/하체`의 대분류로 운동 volume을 기록한다. `Scheduling`
+- [ ] 하루 운동이 끝나면 `가슴/등/하체/어깨/이두/삼두/기타`의 대분류로 운동 volume을 기록한다. `Scheduling`
 - [ ] 한 주가 끝나면 주간 운동 기록을 유저 별로 따로 저장한다. `Scheduling`
 
 ### 회원 별 운동 통계 제공
 - [ ] 주간, 월간 운동 기록, 통계를 제공한다.
   - [ ] 운동 부위 별 volume 주간 통계 제공
-  - `!!! 이후 추가 예정 !!!`
 
 ## 루틴 만들기
 - [ ] 하나의 루틴은 여러개의 `ROUTINE_ELEMENT`로 구성된다.
 - [ ] 회원은 자신만의 루틴을 만들 수 있다.
 
-
-## 운동 커뮤니티 
-( 시간이 되는대로 추가 예정 )
-### 자유 게시판
-- [ ] 카테고리에 상관없는 자유 게시판
+## 운동 커뮤니티
 ### 루틴 공유 게시판
-- [ ] 회원 자신의 루틴을 공유할 수 있다.
-### 오늘 운동 완료 게시판
-- [ ] 회원은 오늘 운동 완료 후 인증샷을 추가할 수 있다.
-- [ ] 원하는 경우 오늘 운동 기록을 공유할 수 있다.
+- [ ] 회원은 자신의 루틴을 공유할 수 있다.
+- [ ] 회원은 다른 회원이 공유한 루틴을 확인하고 댓글과 좋아요를 할 수 있다.
+  - [ ] 다른 회원의 루틴을 저장할 수 있다.
 
 ## ERD
 - ERD는 아직 완성본이 아닙니다.
