+++ src/main/java/com/example/fitnessrecord/community/routinepost/persist/RoutinePost.java
@@ -45,4 +45,5 @@ public class RoutinePost {
 
   private int views;
   private int likes;
+
 }
