+++ src/main/java/com/example/fitnessrecord/domain/training/common/type/BodyPart.java
@@ -8,15 +8,10 @@ import lombok.Getter;
 public enum BodyPart {
   CHEST("가슴(Chest)"),
   BACK("등(Back)"),
+  LEGS("하체(Legs)"),
+  SHOULDER("어깨(Shoulder)"),
   BICEPS("이두(Biceps)"),
   TRICEPS("삼두(Triceps)"),
-  SHOULDER("어깨(Shoulder)"),
-  ARM("팔(Arms)"),
-  ABS("복근(Abs)"),
-  LEGS("하체(Legs)"),
-  QUADS("대퇴 사두(Quads)"),
-  HAMSTRINGS("햄스트링(Hamstrings)"),
-  CALVES("종아리(Calves)"),
   ETC("기타(Etc)");
 
   private final String description;
